msm: camera2: cpp: Remove the reinstallation of cpp_timer
when msm_cpp_do_timeout_work is running, it's possible that the
cpp_timer is pending again. So if we setup this timer in the work
again, it will initialize the timer list entry without any
synchronization. If other context is accessing the global timer
list, it will found one of the node being corrupted. And finally
system crashes. Actually we don't need to reinstall the timer
because the callback function and private data don't change at all.

Change-Id: I51ba8a0a4e575af6b9f2816245e4c5f4a1d99fd1
CRs-Fixed: 637802
Signed-off-by: Figo Wang <figow@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>