Revert msm: clock-krait-8974: Change ref count behavior at boot Set up the refcount of the L2 clock to be equal to the number of online cores. This allows the clients of the CPU/L2 clocks that enable/disable them to be agnostic of the no. of cores that are online.
For example, the CPU idle code can call clk_disable() on the L2 clock from
all the cores' idle path without having to figure out if it's the last
active core. If the core is the last core that's active, the ref count will
go from 1 to 0 and the clock would get disabled.

Also update cpufreq driver behavior to accomodate this change.